# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI will API
  description: |
    AI will - AI音声会話アプリのREST API仕様書

    ## 認証
    JWT Bearer Tokenを使用します。
    ```
    Authorization: Bearer {access_token}
    ```

    ## エラーレスポンス
    すべてのエラーは統一フォーマットで返却されます。
    ```json
    {
      "error": {
        "code": "error_code",
        "message": "エラーメッセージ",
        "details": []
      }
    }
    ```

    ## ページング
    Cursor方式を採用。`cursor`と`limit`パラメータを使用。
    ```json
    {
      "data": [...],
      "pagination": {
        "next_cursor": "eyJpZCI6...",
        "has_more": true
      }
    }
    ```

    ## 冪等性（Idempotency）
    購入系・ジョブ作成系APIでは`Idempotency-Key`ヘッダーが必須です。
    - キー: クライアント生成のUUID
    - 有効期限: 24時間
    - 同一キーで再リクエスト時は同一レスポンスを返却

    ## ストリーミング（SSE）
    会話APIでは Server-Sent Events によるストリーミングをサポートします。
    - エンドポイント: `POST /threads/{id}/messages:stream`
    - イベント: `message_start`, `content_delta`, `message_done`, `error`

    ## 主要なユースケース例
    - **Example 1**: 会話メッセージ送信（`POST /threads/{id}/messages`）
    - **Example 2**: 購入反映・Idempotency-Key付き（`POST /purchases`）
    - **Example 3**: クリップ作成（`POST /clips`）
  version: 1.3.0
  contact:
    name: AI will Support
    email: support@aiwill.app
  license:
    name: Proprietary
    url: https://aiwill.app/terms
servers:
  - url: https://api.aiwill.app/v1
    description: Production
  - url: https://api.staging.aiwill.app/v1
    description: Staging
...
paths:
...
  /clips:
    get:
      tags: [Memory]
      summary: クリップ一覧取得
      description: ユーザーの保存したクリップ（メモリー）一覧を取得します。
      operationId: listClips
      security:
        - BearerAuth: []
      parameters:
        - name: thread_id
          in: query
          schema:
            type: string
            format: uuid
          description: スレッド絞り込み
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClipListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Memory]
      summary: クリップ作成
      description: |
        メッセージを切り抜いてクリップとして保存します。

        ## 用途
        - 思い出に残したい会話をブックマーク
        - 後から振り返りたいシーンを保存
        - タグで整理して検索可能に

        ## 注意事項
        - クリップは作成後に編集不可（title/tags の変更不可）
        - 同一メッセージから複数のクリップを作成可能
        - source_message_id は自分が所有するスレッドのメッセージのみ指定可能
      operationId: createClip
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClipRequest'
            examples:
              date_promise:
                summary: デートの約束を保存
                description: |
                  **Example 3: クリップ作成**

                  キャラクターとの思い出の会話を切り抜いて保存する例。
                  タイトルとタグを付けて、後から検索しやすくします。
                value:
                  source_message_id: msg_550e8400-e29b-41d4-a716-446655440002
                  title: 初めてのデートの約束
                  tags:
                    - デート
                    - 思い出
                    - 水族館
              confession:
                summary: 告白シーンを保存
                value:
                  source_message_id: msg_550e8400-e29b-41d4-a716-446655440099
                  title: あかりの告白
                  tags:
                    - 告白
                    - 感動
                    - 大切な思い出
              no_tags:
                summary: タグなしで保存
                value:
                  source_message_id: msg_550e8400-e29b-41d4-a716-446655440050
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClipResponse'
              examples:
                clip_with_tags:
                  summary: タグ付きクリップ作成成功
                  description: |
                    **Example 3: クリップ作成成功レスポンス**

                    メッセージから切り抜いたクリップが保存され、
                    キャラクター情報と共に返却されます。
                  value:
                    clip:
                      id: clip_550e8400-e29b-41d4-a716-446655440000
                      title: 初めてのデートの約束
                      content: えっと…水族館とか行ってみたいな！イルカのショー見たことある？
                      character:
                        id: char_550e8400-e29b-41d4-a716-446655440001
                        name: 桜井 あかり
                        avatar_url: https://cdn.aiwill.app/characters/char_001/avatar.jpg
                      source_message_id: msg_550e8400-e29b-41d4-a716-446655440002
                      tags:
                        - デート
                        - 思い出
                        - 水族館
                      created_at: '2026-01-31T14:00:05Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /clips/{clip_id}:
    get:
      tags: [Memory]
      summary: クリップ詳細取得
      description: 指定したクリップの詳細を取得します。
      operationId: getClip
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/clipId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClipResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    # PATCH /clips/{clip_id} は削除（v1.1.0）
    # 理由: memory_clips テーブルは updated_at を持たず、作成後の編集は不可とする設計判断
    delete:
      tags: [Memory]
      summary: クリップ削除
      description: 指定したクリップを削除します。
      operationId: deleteClip
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/clipId'
      responses:
        '204':
          description: 削除成功
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ----------------------------------------------------------
  # Safety
  # ----------------------------------------------------------
...
  /privacy/export:
    post:
      tags: [Privacy]
      summary: データエクスポートジョブ作成
      description: |
        ユーザーデータのエクスポートジョブを作成します。

        **冪等性:** `Idempotency-Key`ヘッダー必須
      operationId: createExportJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportJobRequest'
      responses:
        '202':
          description: ジョブ作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: ジョブ進行中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /privacy/export/{job_id}:
    get:
      tags: [Privacy]
      summary: エクスポートジョブ状態確認
      description: エクスポートジョブの状態を確認します。
      operationId: getExportJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /privacy/delete:
    post:
      tags: [Privacy]
      summary: データ削除ジョブ作成
      description: |
        ユーザーデータの削除ジョブを作成します。

        **冪等性:** `Idempotency-Key`ヘッダー必須
      operationId: createDeleteJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeleteJobRequest'
      responses:
        '202':
          description: ジョブ作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteJobResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: ジョブ進行中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /privacy/delete/{job_id}:
    get:
      tags: [Privacy]
      summary: 削除ジョブ状態確認
      description: 削除ジョブの状態を確認します。
      operationId: getDeleteJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteJobResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /privacy/delete/{job_id}/cancel:
    post:
      tags: [Privacy]
      summary: 削除ジョブキャンセル
      description: 猶予期間内の削除ジョブをキャンセルします。
      operationId: cancelDeleteJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: キャンセル成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteJobResponse'
        '400':
          description: キャンセル不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

# ============================================================
# Components
# ============================================================
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
...
    clipId:
      name: clip_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Clip ID
...
    jobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
      description: Job ID
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: ページングカーソル
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 取得件数
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 冪等性キー（UUID）

  # ----------------------------------------------------------
  # Responses
  # ----------------------------------------------------------
  responses:
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_error
              message: リクエストの検証に失敗しました
              details:
                - field: email
                  code: invalid_format
                  message: 有効なメールアドレス形式で入力してください
    UnauthorizedError:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthenticated
              message: 認証が必要です
              details: []
    ForbiddenError:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: forbidden
              message: この操作を実行する権限がありません
              details: []
...
    NotFoundError:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: not_found
              message: リソースが見つかりません
              details: []
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: 次ページのカーソル
        has_more:
          type: boolean
          description: 次ページの有無

    # Auth
...
    Character:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatar_url:
          type: string
          format: uri
...
    Clip:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        content:
          type: string
        character:
          $ref: '#/components/schemas/Character'
        source_message_id:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    CreateClipRequest:
      type: object
      required: [source_message_id]
      properties:
        source_message_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 100
        tags:
          type: array
          items:
            type: string

    # UpdateClipRequest は削除（v1.1.0）
    # 理由: クリップは作成後の編集不可とする設計判断
    ClipResponse:
      type: object
      properties:
        clip:
          $ref: '#/components/schemas/Clip'
    ClipListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Clip'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Safety
...
    JobStatus:
      type: string
      enum: [queued, processing, completed, cancelled, failed]
    JobProgress:
      type: object
      properties:
        total_items:
          type: integer
        processed_items:
          type: integer
        percentage:
          type: integer
    CreateExportJobRequest:
      type: object
      required: [scope]
      properties:
        scope:
          type: string
          enum: [conversations, memories, all]
          description: エクスポート範囲
    ExportJobResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        scope:
          type: string
        progress:
          $ref: '#/components/schemas/JobProgress'
        download_url:
          type: string
          format: uri
          nullable: true
        download_expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    CreateDeleteJobRequest:
      type: object
      required: [scope, confirm]
      properties:
        scope:
          type: string
          enum: [conversations, memories, all]
          description: 削除範囲
        confirm:
          type: boolean
          description: 確認フラグ（true必須）
    DeleteJobResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        scope:
          type: string
        progress:
          $ref: '#/components/schemas/JobProgress'
        grace_period_until:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true